# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Mi API de Servicios",
        "description": "API para gestionar servicios en DynamoDB",
        "version": "1.0.0"
    },
    "paths": {
        "/services": {
            "get": {
                "summary": "Get Services",
                "operationId": "get_services_services_get",
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Authorization"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Service"
                                    },
                                    "title": "Response Get Services Services Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Add Service",
                "operationId": "add_service_services_post",
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Authorization"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NewService"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NewService"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "NewService": {
                "properties": {
                    "service_name": {
                        "type": "string",
                        "title": "Service Name"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status"
                    },
                    "owner": {
                        "type": "string",
                        "title": "Owner"
                    },
                    "latest_change": {
                        "type": "string",
                        "title": "Latest Change"
                    }
                },
                "type": "object",
                "required": [
                    "service_name",
                    "status",
                    "owner",
                    "latest_change"
                ],
                "title": "NewService"
            },
            "Service": {
                "properties": {
                    "service_name": {
                        "type": "string",
                        "title": "Service Name"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status"
                    },
                    "owner": {
                        "type": "string",
                        "title": "Owner"
                    },
                    "latest_change": {
                        "type": "string",
                        "title": "Latest Change"
                    }
                },
                "type": "object",
                "required": [
                    "service_name",
                    "status",
                    "owner",
                    "latest_change"
                ],
                "title": "Service"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}